requirements.txt
Flask
gunicorn
joblib
numpy
pytesseract
Pillow
google-generativeai